// src/app/models/products/mt-Products.ts
export interface IProduct {
    id:   number;
    description:   string;
    rich_description: string;
    image:     string;
    brand:        string;
    price:        number;
    category:     string;
    rating:       string;
    is_featured:  string;
    user_updated: string;
    date_created: string;
    date_updated: string;
    imageUrls: IImages[];
  }

  export interface IImages {
    url: string;
    description: string;
  }

  

// Stripe secret key : sk_test_2v4fYlW3CVfpjbLS1x7vtSfO
// sk_test_2v4fYlW3CVfpjbLS1x7vtSfO
// Product ID: price_1LuscbG9uU7PPnpFTmtL1c9i
// Product ID: price_1Lusi4G9uU7PPnpFlaYvqo4u
// Product ID: price_1LusmDG9uU7PPnpFIhYQcD84